Magix sanity.
