--- src/main/java/emu/grasscutter/command/commands/Broadcast.java
@@ -1,29 +1,0 @@
-package emu.grasscutter.command.commands;
-
-import emu.grasscutter.Grasscutter;
-import emu.grasscutter.command.Command;
-import emu.grasscutter.command.CommandHandler;
-import emu.grasscutter.game.GenshinPlayer;
-
-import java.util.List;
-
-@Command(label = "broadcast", usage = "broadcast <message>",
-        description = "Sends a message to all the players", aliases = {"b"}, permission = "server.broadcast")
-public class Broadcast implements CommandHandler {
-
-    @Override
-    public void execute(GenshinPlayer sender, List<String> args) {
-        if (args.size() < 1) {
-            CommandHandler.sendMessage(sender, "Usage: broadcast <message>");
-            return;
-        }
-
-        String message = String.join(" ", args.subList(0, args.size()));
-
-        for (GenshinPlayer p : Grasscutter.getGameServer().getPlayers().values()) {
-            CommandHandler.sendMessage(p, message);
-        }
-
-        CommandHandler.sendMessage(sender, "Message sent.");
-    }
-}
