+++ src/main/java/emu/grasscutter/command/commands/ReloadCommand.java
@@ -1,0 +1,22 @@
+package emu.grasscutter.command.commands;
+
+import emu.grasscutter.Grasscutter;
+import emu.grasscutter.command.Command;
+import emu.grasscutter.command.CommandHandler;
+import emu.grasscutter.game.GenshinPlayer;
+
+import java.util.List;
+
+@Command(label = "reload", usage = "reload",
+        description = "Reload server config", permission = "server.reload")
+public final class ReloadCommand implements CommandHandler {
+
+    @Override
+    public void execute(GenshinPlayer sender, List<String> args) {
+        CommandHandler.sendMessage(sender, "Reloading config.");
+        Grasscutter.loadConfig();
+        Grasscutter.getGameServer().getGachaManager().load();
+        Grasscutter.getDispatchServer().loadQueries();
+        CommandHandler.sendMessage(sender, "Reload complete.");
+    }
+}
